attempt to rewrite printf to mktemp
When I need to filter contents into a variable as command substitution
and then need to write that variable to stdout for process substitution
if I print the variable with printf or here-string,
all characters are interpreted as raw text instead of utf-8
and some of the multibyte characters break on the other side.
Attempt to write all command substitution variables to mktemp
and cat instead of printf or here-string.